Set up continuous integration
Continuous integration through Travis CI, and code coverage through
CodeCov have been set up. An integration build target has been developed
that can be used by Travis CI and CodeCov to perform continuous
integration and code coverage reporting on the GitHub repository of the
Sol Library.

Support for Clang was attempted, but failed the integration build owing
to a mismatch between Clang and Gcov. This will need to be addressed in
a subsequent issue.

This commit was squashed from the issue/14 branch.