Add unit tests for exception handling module
Unit tests have been prepared for the following interfaces of the
exception handling module:
  * sol_assert()
  * sol_try()
  * sol_erno_now()

The provision of these unit tests has also resulted in the indirect
testing of the following exception handling and unit testing module
interfaces:
  * SOL_TRY
  * SOL_CATCH
  * sol_throw()
  * sol_tsuite_init2()
  * sol_tsuite_term()
  * sol_tsuite_register()
  * sol_tsuite_exec()
  * sol_tsuite_pass()
  * sol_tsuite_total()

Even so, unit tests for the unit testing module need to be incorporated.
Additionally, CLang has been introduced to the development process; this
will allow continuous integration to be done with CLang in addition to
GCC once llvm-cov is incorporated in the development process.

This commit has been squashed from the 'issue/18' branch.