Add unit tests for testing module
Unit tests have been incorporated for the following interfaces of the
unit testing module:
  * sol_tsuite_init()
  * sol_tsuite_init2()
  * sol_tsuite_register()
  * sol_tsuite_pass()
  * sol_tsuite_fail()
  * sol_tsuite_total()
  * sol_tsuite_exec()

These unit tests bring the coverage of the unit testing module to 100%.
A simplified paradigm of defining unit tests has also been introduced,
leading to improved code readability.

Although the unit testing module is designed to be thread-safe by
excluding the use of module-level static global variables, its test
suite is not so; this is acceptable as the test suite is meant to be run
only during the build process, and is **not** a part of the Sol Library
API.

This commit has been squashed from the 'issue/21' branch.