Add logging module
The logging module has been implemented with the following interface
functions/macros:
  * sol_log_open()
  * sol_log_open2()
  * sol_log_close()
  * sol_log_trace()
  * sol_log_debug()
  * sol_log_warn()
  * sol_log_error()
  * sol_log_erno()

In order to implement this interface, a libc compatiblity module has
also been introduced. This module provides a portable means to access
the required libc functions. Since the pointer module makes use of the
libc malloc()/free() functions if available, it has also been refactored
to make use of the newly introduced libc module.

Additionally, the compiler hint module has been enhanced with the
introduction of the sol_tls hint to mark thread local storage variables;
this resulted as a direct necessity of the logging module requiring the
use of a thread local global file stream.

The sol_erno_str() function has been introduced to the exception handling
module to generate the string representation of a sol_erno code; this
function is directly invoked by the sol_log_erno() wrapper macro.

29 unit tests have been added for the logging module, and the features
section of the Readme have been updated with the relevant details.

This commit has been squashed from the 'issue/42' branch.